log_analysis_project/requirements.txt
flask==2.3.2
sqlalchemy==2.0.18
pydantic==2.0.3
python-dotenv==1.0.0
openrouter==0.3.0
requests==2.31.0
pytest==7.4.0
flask-cors==4.0.0 